Initially, bounds are T.untyped
There are assumptions baked in elsewhere that type members behave as
T.untyped until after the resolver. As a result, setting the bounds of
the type members to be top and bottom in the namer where they're first
created leads to problems.

`test/testdata/todo/fixed_ordering.rb` is a good example of this.