Two new, expected errors
I think that it's fine to invent whatever behavior for these, because we
already reported the "don't redefine a ClassOrModule symbol as a
StaticField" earlier in the error output.

Arguably, Sorbet is smart enough to know that the name kind of doesn't
matter: if it's a static field that's not nested under the
`Root::Nested` class or module symbol (e.g. because it was mangled
renamed and the thing is actually `<M <C <U Nested>>$1>` now), then that
should be owned elsewhere.